requirements.txt
fpdf==1.7.2
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
PyMuPDF==1.22.1
